Setup a specific class for all url Fixed adblocking.
